Refactor UserSession to separate module
