--- docs/function/controllers.rst
@@ -1,11 +1,0 @@
-controllers
-===========
-This page describes **controllers** module which can be imported as :code:`vlivepy.controllers`
-
-dumpSession()
--------------
-This function has alias of :func:`vlivepy.dumpSession`
-
-loadSession()
--------------
-This function has alias of :func:`vlivepy.loadSession`
