+++ docs/function/session.rst
@@ -1,0 +1,19 @@
+session
+=======
+This page describes **controllers** module which can be imported as :code:`vlivepy.session`
+
+getUserSession()
+----------------
+.. autofunction:: vlivepy.session.getUserSession
+
+UserSession
+-----------
+This function has alias of :class:`vlivepy.UserSession`
+
+dumpSession()
+-------------
+This function has alias of :func:`vlivepy.dumpSession`
+
+loadSession()
+-------------
+This function has alias of :func:`vlivepy.loadSession`
